Run formatter

+++ src/interfaces/IGauge.sol
@@ -1,12 +1,12 @@
 pragma solidity ^0.8.13;
 
 interface IGauge {
-    function deposit(uint amount) external;
+    function deposit(uint256 amount) external;
     function getReward(address account) external;
-    function notifyRewardAmount(uint amount) external;
-    function withdraw(uint shares) external;
-    function balanceOf(address account) external returns (uint);
-    function voter() external view returns(address);
-    function rewardToken() external view returns(address);
-    function stakingToken() external view returns(address);
+    function notifyRewardAmount(uint256 amount) external;
+    function withdraw(uint256 shares) external;
+    function balanceOf(address account) external returns (uint256);
+    function voter() external view returns (address);
+    function rewardToken() external view returns (address);
+    function stakingToken() external view returns (address);
 }
